refactor: Extract shared types, utils, and pure logic to @opad/libs (#119)
Move platform-agnostic code from client/apps/web to client/libs so both
web and future mobile app can share types, formatters, and business logic.

- Move types: Article, Vocabulary, TokenUsageSummary, etc.
- Move utils: formatDate, formatBytes, formatNumber, formatArticleCount
- Extract getCefrCategory/getCefrLabel from getLevelColor (separate meaning from styling)
- Extract parseErrorResponse, calculatePagination, token usage helpers
- All web imports converted from @/types/*, @/lib/* to @opad/libs
- 356 tests passing, Next.js build green
