+++ client/apps/web/app/usage/__tests__/page.test.tsx
@@ -16,7 +16,7 @@ import React from 'react'
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { render, screen, fireEvent, waitFor } from '@testing-library/react'
 import UsagePage from '../page'
-import { TokenUsageSummary } from '@/types/usage'
+import { TokenUsageSummary } from '@opad/libs'
 
 // Mock modules
 vi.mock('next/navigation', () => ({
