+++ client/apps/web/components/ArticleFilter.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { ArticleStatus } from '@/types/article'
+import { ArticleStatus } from '@opad/libs'
 
 interface ArticleFilterProps {
   selectedStatus?: ArticleStatus
