+++ client/apps/web/components/__tests__/UsageSummary.test.tsx
@@ -15,7 +15,7 @@ import React from 'react'
 import { describe, it, expect } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import UsageSummary from '../UsageSummary'
-import { TokenUsageSummary } from '@/types/usage'
+import { TokenUsageSummary } from '@opad/libs'
 
 describe('UsageSummary', () => {
   const mockSummary: TokenUsageSummary = {
