+++ client/apps/web/package.json
@@ -16,6 +16,7 @@
     "check": "biome check ."
   },
   "dependencies": {
+    "@opad/libs": "workspace:*",
     "@aws-sdk/client-s3": "^3.962.0",
     "@tailwindcss/typography": "^0.5.19",
     "mermaid": "^11.12.2",
