+++ client/libs/index.ts
@@ -1,3 +1,7 @@
-// @opad/shared — shared types and pure utilities
-// Types and utils will be extracted in #119
-export {}
+export type { Article, ArticleStatus, ArticleListResponse, ArticleListFilters, Conjugations, Vocabulary, VocabularyCount } from './types/article'
+export type { TokenUsageSummary, TokenUsageRecord, OperationUsage, DailyUsage } from './types/usage'
+export { formatDate, formatDateShort, formatDateTime, formatBytes, formatNumber, formatArticleCount } from './utils/formatters'
+export { getCefrCategory, getCefrLabel, type CefrCategory } from './utils/cefr'
+export { extractAgentName, formatOperationName, getOperationLabel, formatTokens, formatCost } from './utils/tokenUsage'
+export { parseErrorResponse } from './utils/errorHandling'
+export { calculatePagination, type PaginationInput, type PaginationResult } from './utils/pagination'
